Move splitArray and getWhere into jdbirepo
