+++ dc-cudami-server/dc-cudami-server-backend-jdbi/src/main/java/de/digitalcollections/cudami/server/backend/impl/jdbi/JdbiRepositoryImpl.java
@@ -1,7 +1,7 @@
 package de.digitalcollections.cudami.server.backend.impl.jdbi;
 
-import de.digitalcollections.cudami.server.backend.api.repository.identifiable.IdentifiableRepository;
 import de.digitalcollections.cudami.server.backend.impl.database.AbstractPagingAndSortingRepositoryImpl;
+import de.digitalcollections.cudami.server.backend.impl.jdbi.identifiable.SearchTermTemplates;
 import de.digitalcollections.model.UniqueObject;
 import de.digitalcollections.model.list.filtering.FilterCriterion;
 import de.digitalcollections.model.list.filtering.FilterOperation;
@@ -17,6 +17,7 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Function;
@@ -161,8 +162,10 @@ public abstract class JdbiRepositoryImpl<U extends UniqueObject>
       return;
     }
 
-    // FIXME: modifying pageResponse afterwards should not only change current page but all pages !
-    // seems to be not possible, as other pages base on SQL filtering.... rethink the whole thing...
+    // FIXME: modifying pageResponse afterwards should not only change current page
+    // but all pages !
+    // seems to be not possible, as other pages base on SQL filtering.... rethink
+    // the whole thing...
     for (Map.Entry<String, Function<U, Optional<LocalizedText>>> entry :
         localizedTextFields.entrySet()) {
       String fieldName = entry.getKey();
@@ -207,8 +210,7 @@ public abstract class JdbiRepositoryImpl<U extends UniqueObject>
     if (matchLanguage.find()) {
       // there is a language...
       Locale language = Locale.forLanguageTag(matchLanguage.group(1));
-      List<String> searchTerms =
-          Arrays.asList(IdentifiableRepository.splitToArray((String) filter.getValue()));
+      List<String> searchTerms = Arrays.asList(splitToArray((String) filter.getValue()));
       List<U> filteredContent =
           pageResponse.getContent().parallelStream()
               .filter(
@@ -218,8 +220,7 @@ public abstract class JdbiRepositoryImpl<U extends UniqueObject>
                     if (text == null) {
                       return false;
                     }
-                    List<String> splitText =
-                        Arrays.asList(IdentifiableRepository.splitToArray(text));
+                    List<String> splitText = Arrays.asList(splitToArray(text));
                     return splitText.containsAll(searchTerms);
                   })
               .collect(Collectors.toList());
@@ -281,6 +282,49 @@ public abstract class JdbiRepositoryImpl<U extends UniqueObject>
   protected String getWhereClause(
       FilterCriterion<?> fc, Map<String, Object> argumentMappings, int criterionCount)
       throws IllegalArgumentException, UnsupportedOperationException {
+    Set<String> fieldNames = getLocalizedTextFields().keySet();
+    if (!fieldNames.isEmpty()) {
+      // special handling of localized text fields
+      String regexPatternFieldNames = String.join("|", fieldNames);
+
+      Matcher localizedTextFieldsMatcher =
+          Pattern.compile("^(" + regexPatternFieldNames + ")\\b").matcher(fc.getExpression());
+      if (localizedTextFieldsMatcher.find()) {
+        if (!(fc.getValue() instanceof String)) {
+          throw new IllegalArgumentException(
+              "Value of localized text field expression must be a string!");
+        }
+        String localizedTextFieldName = localizedTextFieldsMatcher.group(1);
+        String value = (String) fc.getValue();
+        switch (fc.getOperation()) {
+          case CONTAINS:
+            if (argumentMappings.containsKey(SearchTermTemplates.ARRAY_CONTAINS.placeholder)) {
+              throw new IllegalArgumentException(
+                  "Filtering by localized text fields value are mutually exclusive!");
+            }
+            argumentMappings.put(
+                SearchTermTemplates.ARRAY_CONTAINS.placeholder, splitToArray(value));
+            return SearchTermTemplates.ARRAY_CONTAINS.renderTemplate(
+                tableAlias, "split_" + localizedTextFieldName);
+          case EQUALS:
+            if (argumentMappings.containsKey(SearchTermTemplates.JSONB_PATH.placeholder)) {
+              throw new IllegalArgumentException(
+                  "Filtering by localized text fields value are mutually exclusive!");
+            }
+            Matcher matchLanguage = Pattern.compile("\\.([\\w_-]+)$").matcher(fc.getExpression());
+            String language =
+                matchLanguage.find() ? "\"%s\"".formatted(matchLanguage.group(1)) : "**";
+            argumentMappings.put(
+                SearchTermTemplates.JSONB_PATH.placeholder, escapeTermForJsonpath(value));
+            return SearchTermTemplates.JSONB_PATH.renderTemplate(
+                tableAlias, localizedTextFieldName, language);
+          default:
+            throw new UnsupportedOperationException(
+                "Filtering by localized text field only supports CONTAINS (to be preferred) or EQUALS operator!");
+        }
+      }
+    }
+
     StringBuilder query = new StringBuilder();
     if (fc != null) {
       FilterOperation filterOperation = fc.getOperation();
@@ -565,8 +609,9 @@ public abstract class JdbiRepositoryImpl<U extends UniqueObject>
   }
 
   /*
-   * if filtering has other target object type than actual (this) repository instance
-   * use this method to rename filter expression names to target table alias and column names
+   * if filtering has other target object type than actual (this) repository
+   * instance use this method to rename filter expression names to target table
+   * alias and column names
    */
   protected void mapFilterExpressionsToOtherTableColumnNames(
       Filtering filtering, AbstractPagingAndSortingRepositoryImpl otherRepository) {
@@ -628,5 +673,40 @@ public abstract class JdbiRepositoryImpl<U extends UniqueObject>
                     .findOne()
                     .get());
     return total;
+  }
+
+  public String[] splitToArray(LocalizedText localizedText) {
+    if (localizedText == null) {
+      return new String[0];
+    }
+    List<String> splitLabels =
+        localizedText.values().stream()
+            .map(text -> splitToArray(text))
+            .flatMap(Arrays::stream)
+            .collect(Collectors.toList());
+    return splitLabels.toArray(new String[splitLabels.size()]);
+  }
+
+  public String[] splitToArray(String term) {
+    term = term.toLowerCase();
+    /*
+     * Remove all characters that are NOT: - space - letter or digit - underscore -
+     * hyphen and remove all standalone hyphens (space hyphen space) (flag `U`
+     * stands for Unicode)
+     */
+    term = term.replaceAll("(?iU)[^\\s\\w_-]|(?<=\\s)-(?=\\s)", "");
+    // Look for words with hyphens to split them too
+    Matcher hyphenWords = Pattern.compile("(?iU)\\b\\w+(-\\w+)+\\b").matcher(term);
+    List<String> result =
+        hyphenWords
+            .results()
+            .collect(
+                ArrayList<String>::new,
+                (list, match) -> list.addAll(Arrays.asList(match.group().split("-+"))),
+                ArrayList::addAll);
+    for (String word : term.trim().split("\\s+")) {
+      result.add(word);
+    }
+    return result.toArray(new String[result.size()]);
   }
 }
